Move local dependencies to the new package
