+++ buildSrc/src/main/kotlin/io/spine/dependency/local/McJava.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.internal.dependency.spine
+package io.spine.dependency.local
 
 /**
  * Dependencies on Spine Model Compiler for Java.
